Move traits to flipperzero::io
